ENH: use unique ID to write params of simulation
Generate a unique int to identify simulations in
uncertainty analysis and pass it in params dict.
If the sim is run by MPI, then write the param
file with the calculated error.

Also fix up printing of status messages to screen
such that they are skipped if run by MPI.